+++ app/src/Containers/index.ts
@@ -4,8 +4,8 @@
 
 export { default as EventSearchContainer } from './EventSearchContainer'
 export { default as EventDetailContainer } from './EventDetailContainer'
-export { default as ListContainer } from './ListContainer'
-export { default as ProjectDetailContainer } from './ProjectDetailContainer'
+export { default as ListContainer } from '../NativeBase/Containers/ListContainer'
+export { default as ProjectDetailContainer } from '../NativeBase/Containers/ProjectDetailContainer'
 export { default as ProjectRegisterInterestContainer } from './ProjectRegisterInterestContainer'
 export { default as StartupContainer } from './StartupContainer'
 export { default as WelcomeContainer } from './WelcomeContainer'