+++ app/src/NativeBase/Components/FreeSearchBar.tsx
@@ -2,9 +2,9 @@
  * @file Text input for searching.
  */
 
-import { Box, Icon, Input } from 'native-base'
+import { Box, Input } from 'native-base'
 import React, { FC, useState } from 'react'
-import MaterialIcons from 'react-native-vector-icons/MaterialIcons'
+import SearchIconHighlighted from './SearchIconHighlighted'
 
 export interface FreeSearchBarProps {
   handleChangeText?: (updatedText: string) => void
@@ -44,16 +44,7 @@ const FreeSearchBar: FC<FreeSearchBarProps> = ({
       <Input
         accessibilityLabel="Search for text"
         height="12"
-        InputLeftElement={
-          <Icon
-            as={MaterialIcons}
-            _light={{ color: 'accentPurple.100' }}
-            _dark={{ color: 'white' }}
-            ml="2"
-            name="search"
-            size={6}
-          />
-        }
+        InputLeftElement={<SearchIconHighlighted />}
         lineHeight="md"
         marginBottom={marginBottom ?? '4'}
         marginTop={marginTop ?? '0'}