+++ app/src/NativeBase/Containers/index.ts
@@ -3,4 +3,8 @@
  */
 
 export { default as ProfileContainer } from './ProfileContainer'
+export { default as ProjectDetailContainer } from './ProjectDetailContainer'
+export { default as ProjectSearchContainer } from './ProjectSearchContainer'
+export { default as SearchResultsContainer } from './SearchResultsContainer'
 export { default as VerticalStackContainer } from './VerticalStackContainer'
+export { default as WebViewContainer } from './WebViewContainer'